tests: test 'capabilities.rag_audit'
  ... via 'tests/unit/capabilities/test_rag_audit_caps.py'
